Create readme.md
A brief explanation of the approach I have taken for this task:
Initially my aim was to break down this task into small pieces that would serve as a means of organising my time and tracking progress more effectively.
The first goal was to create a simple web page that would serve as a means of displaying the user input box for entry of location and the results returned from Foursquare database. Subsequent checkpoints included successfully returning a list of venues from Foursquare and then implementing a simple display of these results. I wanted to ensure that there would be at least some demonstrable functionality before focusing on the look and feel of the application.
It was also my intention to abstract the code into logical divisions and not have all of the CSS and JavaScript contained within the html file.
There are a number of things missing that I would have liked to have completed given more time - neater code and a more interesting way of displaying the results to name a few.

The repository can be found at: github.com/JakeBC/ANDigital
A deployed example can be found at: jakebarclay.com